refactor: extract repo metadata utils from composable
and reuse instead of duplicating for Likes Leaderboard

Technically there's one behavioural change here: on the Likes Leaderboard, previously if fetching
from a non-GitHub repo, the stars count would always be null and not get rendered, but now for some
providers we will show a stars count of 0 instead of null. This is fine since no packages in the top
10 today have non-GitHub repos.
